+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [NEXT]
+- 
+
 ## [1.7.0] – 2026-07-08
 - Separate header from body claims in trust-statement so that header claims and body claims can be distinguished and validated independently
 - JWKSet no longer used for signature checks for jwt via DidDocument, instead use the JWK from keys directly.