+++ CHANGELOG.md
@@ -5,8 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [NEXT] – next
+## [NEXT]
+- 
+
+## [1.7.0] – 2026-07-08
 - Separate header from body claims in trust-statement so that header claims and body claims can be distinguished and validated independently
+- JWKSet no longer used for signature checks for jwt via DidDocument, instead use the JWK from keys directly.
 
 ## [1.6.2] – 2026-06-03
 - Bump didresolver to 2.8.0 for security updates.